src/workflow/replacement_processor.py
"""替换处理器模块

负责处理视频中的人物、背景和商品替换逻辑。
"""

import os
import cv2
import numpy as np
from typing import Dict, Any, List, Optional
from pathlib import Path

from ..utils.logger import get_logger
from ..utils.file_utils import ensure_directory, safe_copy

logger = get_logger(__name__)


class ReplacementProcessor:
    """替换处理器类
    
    提供视频中人物、背景和商品替换的核心功能。
    """
    
    def __init__(self):
        """初始化替换处理器"""
        logger.info("替换处理器已初始化")
    
    def detect_objects(self, image_path: str, detection_type: str, config: Dict[str, Any]) -> Optional[Dict[str, Any]]:
        """检测图像中的对象
        
        Args:
            image_path: 图像路径
            detection_type: 检测类型 ('person', 'background', 'product')
            config: 检测配置参数
            
        Returns:
            检测结果，包含边界框、掩码等信息
        """
        logger.info(f"检测图像中的{ detection_type }: {image_path}")
        
        # 这里应该实现实际的对象检测逻辑
        # 简化实现，返回一个示例结果
        try:
            image = cv2.imread(image_path)
            if image is None:
                logger.error(f"无法读取图像: {image_path}")
                return None
            
            height, width = image.shape[:2]
            
            # 根据不同的检测类型返回不同的示例结果
            if detection_type == 'person':
                confidence_threshold = config.get('confidence_threshold', 0.7)
                return {
                    'detected': True,
                    'bbox': [width * 0.3, height * 0.2, width * 0.5, height * 0.7],  # [x, y, w, h]
                    'confidence': 0.9,
                    'keypoints': []  # 可以添加人体关键点信息
                }
            elif detection_type == 'product':
                return {
                    'detected': True,
                    'bbox': [width * 0.5, height * 0.6, width * 0.2, height * 0.3],
                    'confidence': 0.85
                }
            elif detection_type == 'background':
                # 背景检测通常是生成整个背景的掩码
                return {
                    'detected': True,
                    'mask': np.ones((height, width), dtype=np.uint8) * 255  # 简化的掩码
                }
            
            return None
            
        except Exception as e:
            logger.error(f"对象检测失败: {str(e)}")
            return None
    
    def create_mask(self, image_path: str, detection_result: Dict[str, Any], mask_type: str) -> Optional[np.ndarray]:
        """创建对象掩码
        
        Args:
            image_path: 图像路径
            detection_result: 检测结果
            mask_type: 掩码类型
            
        Returns:
            掩码数组
        """
        try:
            image = cv2.imread(image_path)
            if image is None:
                logger.error(f"无法读取图像: {image_path}")
                return None
            
            height, width = image.shape[:2]
            mask = np.zeros((height, width), dtype=np.uint8)
            
            if mask_type == 'person' and detection_result.get('detected', False):
                # 创建人物掩码
                bbox = detection_result.get('bbox', [])
                if len(bbox) == 4:
                    x, y, w, h = map(int, bbox)
                    cv2.rectangle(mask, (x, y), (x + w, y + h), 255, -1)
            
            elif mask_type == 'background' and detection_result.get('detected', False):
                # 创建背景掩码（通常是反向的）
                if 'mask' in detection_result:
                    mask = detection_result['mask']
                else:
                    # 如果没有提供掩码，使用默认的背景掩码
                    mask = np.ones((height, width), dtype=np.uint8) * 255
                    # 这里可以添加前景保护逻辑
            
            elif mask_type == 'product' and detection_result.get('detected', False):
                # 创建商品掩码
                bbox = detection_result.get('bbox', [])
                if len(bbox) == 4:
                    x, y, w, h = map(int, bbox)
                    cv2.rectangle(mask, (x, y), (x + w, y + h), 255, -1)
            
            return mask
            
        except Exception as e:
            logger.error(f"创建掩码失败: {str(e)}")
            return None
    
    def apply_replacement(self, image_path: str, mask: np.ndarray, replacement_config: Dict[str, Any], output_path: str) -> bool:
        """应用替换
        
        Args:
            image_path: 原始图像路径
            mask: 替换掩码
            replacement_config: 替换配置
            output_path: 输出图像路径
            
        Returns:
            是否替换成功
        """
        try:
            image = cv2.imread(image_path)
            if image is None:
                logger.error(f"无法读取图像: {image_path}")
                return False
            
            # 确保输出目录存在
            ensure_directory(os.path.dirname(output_path))
            
            # 这里应该实现实际的替换逻辑
            # 简化实现，根据替换类型做不同处理
            
            # 在实际应用中，这里会调用AI模型或其他方法来生成替换内容
            # 这里只是一个占位符
            if replacement_config.get('replacement_image'):
                # 如果提供了替换图像，使用它
                replacement_image = cv2.imread(replacement_config['replacement_image'])
                if replacement_image is not None:
                    # 调整替换图像大小以匹配原始图像
                    replacement_image = cv2.resize(replacement_image, (image.shape[1], image.shape[0]))
                    # 应用掩码
                    result = cv2.bitwise_and(image, image, mask=cv2.bitwise_not(mask))
                    result += cv2.bitwise_and(replacement_image, replacement_image, mask=mask)
                    cv2.imwrite(output_path, result)
                    return True
            
            # 简化处理，直接保存原始图像
            cv2.imwrite(output_path, image)
            logger.warning("替换功能尚未完全实现，返回原始图像")
            return True
            
        except Exception as e:
            logger.error(f"应用替换失败: {str(e)}")
            return False
    
    def process_frame(self, frame_path: str, replacement_config: Dict[str, Any], output_dir: str) -> Optional[str]:
        """处理单个帧
        
        Args:
            frame_path: 帧路径
            replacement_config: 替换配置
            output_dir: 输出目录
            
        Returns:
            处理后的帧路径
        """
        logger.info(f"处理帧: {frame_path}")
        
        # 确保输出目录存在
        ensure_directory(output_dir)
        
        # 生成输出路径
        frame_name = os.path.basename(frame_path)
        output_path = os.path.join(output_dir, frame_name)
        
        # 检查是否需要进行替换
        has_replacement = False
        replacement_types = []
        
        if replacement_config.get('person', {}).get('enabled', False):
            has_replacement = True
            replacement_types.append('person')
        
        if replacement_config.get('background', {}).get('enabled', False):
            has_replacement = True
            replacement_types.append('background')
        
        if replacement_config.get('product', {}).get('enabled', False):
            has_replacement = True
            replacement_types.append('product')
        
        # 如果没有替换需求，直接复制原帧
        if not has_replacement:
            safe_copy(frame_path, output_path)
            return output_path
        
        # 对每种替换类型进行处理
        current_image = frame_path
        for replace_type in replacement_types:
            config = replacement_config.get(replace_type, {})
            
            # 检测对象
            detection_result = self.detect_objects(current_image, replace_type, config)
            if not detection_result or not detection_result.get('detected', False):
                logger.warning(f"在帧中未检测到{replace_type}: {current_image}")
                continue
            
            # 创建掩码
            mask = self.create_mask(current_image, detection_result, replace_type)
            if mask is None:
                logger.error(f"无法创建{replace_type}掩码")
                continue
            
            # 应用替换
            temp_output = os.path.join(output_dir, f"temp_{replace_type}_{frame_name}")
            if self.apply_replacement(current_image, mask, config, temp_output):
                current_image = temp_output
            else:
                logger.error(f"应用{replace_type}替换失败")
        
        # 最终输出
        safe_copy(current_image, output_path)
        
        # 清理临时文件
        for replace_type in replacement_types:
            temp_output = os.path.join(output_dir, f"temp_{replace_type}_{frame_name}")
            if os.path.exists(temp_output) and temp_output != output_path:
                try:
                    os.remove(temp_output)
                except Exception as e:
                    logger.warning(f"无法删除临时文件: {str(e)}")
        
        logger.info(f"帧处理完成: {output_path}")
        return output_path
    
    def process_frames_batch(self, frame_paths: List[str], replacement_config: Dict[str, Any], output_dir: str) -> List[str]:
        """批量处理帧
        
        Args:
            frame_paths: 帧路径列表
            replacement_config: 替换配置
            output_dir: 输出目录
            
        Returns:
            处理后的帧路径列表
        """
        logger.info(f"开始批量处理 {len(frame_paths)} 帧")
        
        processed_paths = []
        for i, frame_path in enumerate(frame_paths):
            processed_path = self.process_frame(frame_path, replacement_config, output_dir)
            if processed_path:
                processed_paths.append(processed_path)
            
            # 记录进度
            if (i + 1) % 10 == 0 or (i + 1) == len(frame_paths):
                progress = (i + 1) / len(frame_paths) * 100
                logger.info(f"帧处理进度: {i + 1}/{len(frame_paths)} ({progress:.1f}%)")
        
        logger.info(f"批量处理完成，成功处理 {len(processed_paths)} 帧")
        return processed_paths


# 创建全局实例
replacement_processor = ReplacementProcessor()