requirements.txt
# 基础依赖
opencv-python==4.8.0.74
numpy==1.24.3
pillow==9.5.0
requests==2.31.0
pydantic==2.0.2
python-dotenv==1.0.0
loguru==0.7.0
tqdm==4.65.0

# 替换功能所需的额外依赖
torch==2.0.1
torchvision==0.15.2
scikit-image==0.21.0
scikit-learn==1.3.0
matplotlib==3.7.2
imageio==2.31.1

# 开发依赖
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.9.1
flake8==6.1.0
isort==5.12.0